Minor bugfix in case of non existing highscores, removed warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2377 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/three_strikes_battle.cpp
@@ -129,9 +129,9 @@ void ThreeStrikesBattle::updateKartRanks()
         for( unsigned int n = 0; n < NUM_KARTS-1; ++n )
         {
             const int this_karts_time = m_kart[karts_list[n]]->hasFinishedRace() ?
-                m_kart[karts_list[n]]->getFinishTime() : TimedRace::getTime();
+                (int)m_kart[karts_list[n]]->getFinishTime() : (int)TimedRace::getTime();
             const int next_karts_time = m_kart[karts_list[n+1]]->hasFinishedRace() ?
-                m_kart[karts_list[n+1]]->getFinishTime() : TimedRace::getTime();
+                (int)m_kart[karts_list[n+1]]->getFinishTime() : (int)TimedRace::getTime();
 
             bool swap = false;
